pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "third-arm"
version = "0.1.0"
description = "Modular supernumerary arm — intent-driven handover service (Stage 1 desktop-first)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "pyyaml>=6.0.1",
    "mcap>=1.1.1",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",         # async HTTP client (future integrations)
    "structlog>=24.1.0",     # structured logging
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
third-arm-dev    = "third_arm.cli.run_dev:main"
third-arm-export = "third_arm.cli.export_bundle:main"
third-arm-replay = "third_arm.cli.replay_session:main"

[tool.hatch.build.targets.wheel]
packages = ["src/third_arm"]

[tool.pytest.ini_options]
testpaths    = ["tests"]
asyncio_mode = "auto"
addopts      = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
strict         = false
ignore_missing_imports = true
